+++ misc/mess01.R
@@ -12,11 +12,11 @@ inv_tau_z <- function(tau) {
 
 tau_ci <- function(tau, n, ci = 0.95) {
   z <- qnorm((1 - ci) /2, lower.tail = FALSE)
-  se_tau_z <- sqrt(0.437/(n-4))
+  var_tau_z <- sqrt(0.437/(n-4))
   tau_z <- tau_z(tau)
   tau_z_ci <- c(
-    lower = tau_z - z * se_tau_z, 
-    upper = tau_z + z * se_tau_z
+    lower = tau_z - z * var_tau_z, 
+    upper = tau_z + z * var_tau_z
   )
   
   list(
@@ -24,13 +24,12 @@ tau_ci <- function(tau, n, ci = 0.95) {
     tau_ci = c(lower = inv_tau_z(tau_z_ci[1]), upper = inv_tau_z(tau_z_ci[2])),
     tau_z = tau_z,
     tau_z_ci = tau_z_ci,
-    se = se_tau_z,
-    z = z
+    var_tau_z = var_tau_z
   )
   
 }
 
-tau_ci(0.6, 20)
+tau_ci(0.3684211, 20)
 tau_z(0.6)