+++ NEWS.md
@@ -1,11 +1,9 @@
 # scan 0.55.x
 
 
-- `tau_u(), correted_tau()`: corrected a wrong calculation of the cotinuity correction when values where lower in phase B.
+## New features
 
 - `export()`: New export for power analysis output.
-
-
 - `select_cases()`: Allow for a selection based on object names (like in substitute).
 
 ```r
@@ -35,6 +33,11 @@ readODS::read_ods("filename.ods") |>
 as.data.frame(exampleABC) |> readODS::write_ods("filename.xlsx")
 as.data.frame(exampleABC) |> openxlsx::write.xlsx("filename.xlsx")
 ```
+
+## Bug fixes
+
+- `tau_u(), correted_tau()`: corrected a wrong calculation of the continuity correction when values where lower in phase B.
+- `tau_u()`: Implemented a new method for calculating confidence intervals based on Fisher-Z transformations (see Long, J. D., & Cliff, N. (1997). Confidence intervals for Kendall’s tau. British Journal of Mathematical and Statistical Psychology, 50(1), 31–41. <https://doi.org/10.1111/j.2044-8317.1997.tb01100.x>)
 
 # scan 0.55